pyproject.toml
[build-system]
requires = ["setuptools>=61.0", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "youtube-intelligence-mcp"
version = "1.0.0"
description = "YouTube Intelligence MCP Server for AI-powered video analysis"
authors = [{ name = "YouTube Intelligence Team" }]
license = { text = "MIT" }
readme = "README.md"
requires-python = ">=3.9"
classifiers = [
    "Development Status :: 4 - Beta",
    "Intended Audience :: Developers",
    "License :: OSI Approved :: MIT License",
    "Programming Language :: Python :: 3",
    "Programming Language :: Python :: 3.9",
    "Programming Language :: Python :: 3.10",
    "Programming Language :: Python :: 3.11",
]
dependencies = [
    "mcp>=0.9.0",
    "google-api-python-client==2.108.0",
    "pandas==2.1.3",
    "numpy==1.24.3",
    "nltk==3.8.1",
    "textblob==0.17.1",
    "anthropic==0.7.7",
    "firebase-admin==6.2.0",
    "python-dotenv==1.0.0",
    "pydantic==2.5.0",
    "aiofiles==23.2.1",
    "psutil==5.9.6",
]

[project.scripts]
youtube-intelligence-mcp = "server.__main__:main"

[tool.setuptools.packages.find]
include = ["*"]
exclude = ["tests*", "*.ipynb_checkpoints*"]
